Rename and simplify user code package
