--- src/main/java/com/team766/EntryPoint.java
@@ -1,9 +1,0 @@
-package com.team766;
-
-import com.team766.hal.MyRobot;
-
-public class EntryPoint {
-	public static MyRobot createRobot() {
-		return new com.team766.frc2022.Robot();
-	}
-}
