+++ src/main/java/com/team766/robot/Robot.java
@@ -1,0 +1,13 @@
+package com.team766.robot;
+
+import com.team766.robot.mechanisms.*;
+
+public class Robot {
+	// Declare mechanisms here
+	
+
+	public static void robotInit() {
+		// Initialize mechanisms here
+		
+	}
+}
