+++ src/main/java/com/team766/robot/procedures/DoNothing.java
@@ -1,4 +1,4 @@
-package com.team766.frc2022.procedures;
+package com.team766.robot.procedures;
 
 import com.team766.framework.Context;
 import com.team766.framework.Procedure;
